refactor(parser): clean up structure
