+++ package-parser/package_parser/cli/_run_annotations.py
@@ -1,9 +1,9 @@
 import json
 from pathlib import Path
 
-from package_parser.model.annotations import AnnotationStore
-from package_parser.model.api import API
-from package_parser.model.usages import UsageCountStore
+from package_parser.processing.annotations.model import AnnotationStore
+from package_parser.processing.api.model.api import API
+from package_parser.processing.usages.model import UsageCountStore
 from package_parser.processing.annotations import generate_annotations
 from package_parser.utils import ensure_file_exists
