+++ package-parser/package_parser/processing/annotations/_usages_preprocessor.py
@@ -1,5 +1,5 @@
-from package_parser.model.api import API
-from package_parser.model.usages import UsageCountStore
+from package_parser.processing.api.model.api import API
+from package_parser.processing.usages.model import UsageCountStore
 from package_parser.utils import parent_id
