+++ package-parser/package_parser/processing/api/documentation/__init__.py
@@ -1,3 +1,4 @@
-from ._AbstractDocumentationParsingStrategy import ClassDocumentation, FunctionDocumentation, ParameterDocumentation
+from ._AbstractDocumentationParser import AbstractDocumentationParser, ClassDocumentation, FunctionDocumentation, \
+    ParameterDocumentation
 from ._NumpyDocParser import NumpyDocParser
 from ._get_full_docstring import get_full_docstring
