+++ package-parser/tests/processing/api/test_boundaries.py
@@ -1,5 +1,5 @@
 import pytest
-from package_parser.model.api._types import BoundaryType
+from package_parser.processing.api.model.api import BoundaryType
 
 
 @pytest.mark.parametrize(
