+++ package-parser/tests/processing/api/test_enums.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 import pytest
-from package_parser.model.api._types import EnumType
+from package_parser.processing.api.model.api import EnumType
 
 
 @pytest.mark.parametrize(
