+++ package-parser/tests/processing/dependencies/test_get_dependency.py
@@ -1,5 +1,5 @@
 import spacy
-from package_parser.model.api import (
+from package_parser.processing.api.model.api import (
     Action,
     Condition,
     Dependency,
